refactor: move ports to application dir
