+++ internal/application/workflow/abort.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/lmtani/pumbaa/internal/application"
-	"github.com/lmtani/pumbaa/internal/domain/ports"
+	"github.com/lmtani/pumbaa/internal/application/ports"
 	workflow2 "github.com/lmtani/pumbaa/internal/domain/workflow"
 )
