+++ internal/application/workflow/outputs.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/lmtani/pumbaa/internal/application"
-	"github.com/lmtani/pumbaa/internal/domain/ports"
+	"github.com/lmtani/pumbaa/internal/application/ports"
 )
 
 // OutputsUseCase handles workflow outputs retrieval.
