+++ internal/infrastructure/cloudlogging/repository.go
@@ -10,7 +10,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/lmtani/pumbaa/internal/domain/ports"
+	"github.com/lmtani/pumbaa/internal/application/ports"
 )
 
 // CloudLoggingRepository implements BatchLogsRepository using Google Cloud Logging.
