+++ internal/infrastructure/metrics/tsv_writer.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/lmtani/pumbaa/internal/domain/ports"
+	"github.com/lmtani/pumbaa/internal/application/ports"
 	"github.com/lmtani/pumbaa/internal/domain/workflow"
 )
