+++ internal/infrastructure/recommendation/debug_writer.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/lmtani/pumbaa/internal/domain/ports"
+	"github.com/lmtani/pumbaa/internal/application/ports"
 )
 
 // FileDebugWriter implements LLMDebugWriter by writing to a file.
