+++ internal/infrastructure/storage/file_size_cache.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"sync"
 
-	"github.com/lmtani/pumbaa/internal/domain/ports"
+	"github.com/lmtani/pumbaa/internal/application/ports"
 )
 
 // FileSizeCache provides thread-safe caching of file sizes with persistent storage.
