+++ internal/interfaces/cli/handler/debug.go
@@ -7,9 +7,9 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/urfave/cli/v2"
 
+	"github.com/lmtani/pumbaa/internal/application/ports"
 	workflowapp "github.com/lmtani/pumbaa/internal/application/workflow"
 	"github.com/lmtani/pumbaa/internal/config"
-	"github.com/lmtani/pumbaa/internal/domain/ports"
 	"github.com/lmtani/pumbaa/internal/infrastructure/chat/agent/tools"
 	"github.com/lmtani/pumbaa/internal/infrastructure/chat/llm"
 	cromwellclient "github.com/lmtani/pumbaa/internal/infrastructure/cromwell"
